Add calver as a build requirement
calver is actually a build requirement of trove-classifiers, but it goes
undetected by pip_find_builddeps.py because of [1].

This causes disconnected builds that rely on explicit requirements
listings (like those done with Cachito) to fail. By adding it explicitly
those builds should be fixed.

Future runs of pip_find_builddeps.py will preserve this manual addition
as long as the --append argument is passed to it.

Related: RELDEL-6094

[1] https://github.com/pypa/pip/issues/7863